requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# HTTP client and WebSocket support
httpx==0.25.2
websockets==12.0

# Database and Supabase
supabase==2.3.0
asyncpg==0.29.0

# Audio processing
speechbrain==0.5.16
torchaudio==2.1.1
torch==2.1.1
numpy==1.24.4
librosa==0.10.1
ffmpeg-python==0.2.0

# Observability and logging
structlog==23.2.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-httpx==0.42b0

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
testcontainers==3.7.1
httpx==0.25.2

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1